Restored test code and install script ammendment
